+++ Source/Classes/GLIContext.m
@@ -12,7 +12,6 @@
 @interface GLIContext ()
 {
     EAGLContext *_glContext;
-    //pthread_mutex_t _lock;
 }
 
 @end
@@ -38,11 +37,6 @@
 {
     if (self = [super init])
     {
-//        pthread_mutexattr_t attr;
-//        pthread_mutexattr_init(&attr);
-//        pthread_mutexattr_settype(&attr, PTHREAD_MUTEX_RECURSIVE);
-//        pthread_mutex_init(&_lock, &attr);
-
         _glContext = [[EAGLContext alloc] initWithAPI:kEAGLRenderingAPIOpenGLES2];
 
         [[NSNotificationCenter defaultCenter] addObserver:self selector:@selector(applicationWillResignActive:) name:UIApplicationWillResignActiveNotification object:nil];
@@ -54,33 +48,12 @@
 
 - (void)applicationWillResignActive:(NSNotification *)notification
 {
-    //[self lock];
     glFinish();
-    //[self unlock];
 }
 
 - (void)applicationDidEnterBackground:(NSNotification *)notification
 {
-    //[self lock];
     glFinish();
-    //[self unlock];
 }
-
-/*
-- (void)lock
-{
-    pthread_mutex_lock(&_lock);
-}
-
-- (int)tryLock
-{
-    return pthread_mutex_trylock(&_lock);
-}
-
-- (void)unlock
-{
-    pthread_mutex_unlock(&_lock);
-}
-//*/
 
 @end